fix: cloudvolume can read/write fpzip but something is weird with it
